requirements.txt
asgiref==3.11.0
black==25.12.0
click==8.3.1
Django==6.0.1
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-stubs==3.16.7
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
pillow==12.1.0
platformdirs==4.5.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytokens==0.3.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
typing_extensions==4.15.0
urllib3==2.6.3
